重构plugin_manager添加获取系统插件路径的方法
